Port v4_s1_v4_ll: three robots, the scene V4 authored, and a way to look at it
The task itself was the easy half. What took the work was everything that has
to be true before an episode is worth running.

Robots. All three imported from their own upstream URDFs through BEHAVIOR's
official importer. The Crazyflie needed five separate things, each found by
loading it rather than by reading:

  articulation root was None       use_holonomic_joints, which also gives the
                                   altitude DOF a drone needs -- the
                                   non-holonomic branch of
                                   _get_robot_pose_from_2d_pose forces z to 0.0
  no base controller group         holonomic_base + locomotion
  no end effector, could not grasp a zero-joint arm; the body is the suction face
  reset_joint_pos was None         default_joint_pos, one per virtual joint
  arm_control_idx empty -> float32 dtype=th.int (robot.py; the sibling of the
                                   gripper fix, and the one I said I would not
                                   touch until it fired -- it fired)

And then it loaded, was correctly placed, correctly sized, and rendered
nothing. OmniGibson hides the eef link of every manipulation robot on principle,
because on a normal arm that link is a massless frame like ur_arm_tool0 --
naming the body as the end effector hid the whole drone. It now has its own
bottom_suction_mount 1.5 cm below the belly, which is what V4's own wrapper
adds and for the same reason.

Scene. The instance pins the box to V4's coordinate rather than sampling one,
and reproduces the scene V4 authored: 84 objects removed (19 pictures, 16 light
switches, 10 doors, the sofa, five armchairs), two moved, the box at 8 g. The
filter list is a separate USD layer the staging scenes reference -- easy to miss
by reading only their own edit block, which is how I first counted 8 removals
instead of 89. Beds, the coffee table, shelf and fridge stay: they are BDDL
supports. The earlier NO_SPACE_AROUND_TARGET failures named
electric_switch_wseglt_8 and picture_zsirgc_0, both on that list.

Layouts carry scale now, because size is part of a layout and not a property of
a model. V4 stages its three robots within half a metre of each other, which
only fits because it shrinks them to 0.5/0.7/0.6. Loaded full size two of them
overlapped, place_robots relocated them by 0.8 m and 1.9 m, and the relocation
shoved the box 1.4 m across the room.

inspect_scene loads all of it and stops -- no agents, no LLM, no plan loop --
and prints where each robot was asked to be against where it ended up. It found
three defects in its first three runs, none of which cost an episode: a YAML
generator that split on "robots:" and matched include_robots:, prepare_robots
asserting every robot has a gripper, and the drone.

+++ OmniGibson/omnigibson/robots/robot.py
@@ -2813,7 +2813,14 @@ class Robot(USDObject, GymObservable):
         """
         assert self.is_manipulation
         idxs = {
-            arm: th.tensor([list(self.joints.keys()).index(name) for name in self.arm_joint_names[arm]])
+            # dtype is explicit for the same reason as gripper_control_idx
+            # below: an arm with no joints of its own is legitimate -- a suction
+            # cup fixed to a drone's belly is an end effector with no chain --
+            # and th.tensor([]) is float32, which cannot index a tensor.
+            arm: th.tensor(
+                [list(self.joints.keys()).index(name) for name in self.arm_joint_names[arm]],
+                dtype=th.int,
+            )
             for arm in self.arm_names
         }
         if self._definition.manipulation and self._definition.manipulation.add_combined_arm_control_idx:
